+++ ctrl/snap_virtio_net_ctrl.h
@@ -45,10 +45,10 @@ struct snap_virtio_net_ctrl_queue {
 };
 
 struct snap_virtio_net_ctrl_lm_cbs {
-    unsigned (*get_internal_state_size)(void *cb_ctx);
-    int (*get_internal_state)(void *cb_ctx, void *buf, unsigned len);
-    unsigned (*dump_internal_state)(void *cb_ctx, void *buff, unsigned len);
-    int (*set_internal_state)(void *cb_ctx, void *buf, unsigned len);
+	size_t (*get_internal_state_size)(void *cb_ctx);
+	int (*get_internal_state)(void *cb_ctx, void *buf, size_t len);
+	size_t (*dump_internal_state)(void *cb_ctx, void *buff, size_t len);
+	int (*set_internal_state)(void *cb_ctx, void *buf, size_t len);
 };
 
 struct snap_virtio_net_ctrl_attr {
@@ -64,7 +64,7 @@ struct snap_virtio_net_ctrl {
 static inline struct snap_virtio_net_ctrl*
 to_net_ctrl(struct snap_virtio_ctrl *vctrl)
 {
-       return container_of(vctrl, struct snap_virtio_net_ctrl, common);
+	return container_of(vctrl, struct snap_virtio_net_ctrl, common);
 }
 
 struct snap_virtio_net_ctrl*
